pyproject.toml
[tool.poetry]
name = "rag-backend"
version = "0.1.0"
description = ""
authors = ["JeanGargano <159099449+JeanGargano@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
chromadb = "^0.5.5"
onnxruntime = "^1.18.1"
pydantic-settings = "^2.4.0"
pytest_plugins = ['pytest_asyncio']
protobuf = ">=3.20.0,<3.21.0"
openai = "^1.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
